pyproject.toml
[tool.poetry]
name = "research"
version = "0.1.0"
description = ""
authors = ["da.sadovenko <sadovenkoda@dialog-regions.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"
aiokafka = "^0.8.0"
pandas = "^2.0.2"
httpx = "^0.23.3"
beautifulsoup4 = "^4.11.2"
asyncio-cache = "^0.0.1"
requests-html = "^0.10.0"
openpyxl = "3.1.2"
pydantic = "^2.0.2"
xmltodict = "^0.13.0"
lxml = "4.9.3"
dask = "2023.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
